% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumour.R
\name{rumour}
\alias{rumour}
\title{Generate a rumour}
\usage{
rumour(suspect, weapon, room)
}
\arguments{
\item{suspect}{the suspect}

\item{weapon}{a weapon}

\item{room}{the room}
}
\value{
a rumour, who did it, in what room, with what weapon
}
\description{
Generate a rumour
}
