% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumour.R
\name{whodunnit}
\alias{whodunnit}
\title{Whodunnit}
\usage{
whodunnit(x = 1)
}
\arguments{
\item{x}{how many rumours do you want returned}
}
\value{
x number of rumours
}
\description{
Whodunnit
}
\examples{
whodunnit()
whodunnit(5)
}
