% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumour.R
\name{rumour}
\alias{rumour}
\title{Generate a rumour}
\usage{
rumour(suspect, weapon, room)
}
\arguments{
\item{suspect}{string value of the suspect}

\item{weapon}{string value of the weapon}

\item{room}{string value of the room}
}
\value{
a rumour based on who did the deed, in which room and with what weapon.
}
\description{
Generate a rumour
}
