% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumour.R
\name{whodunnit}
\alias{whodunnit}
\title{Whodunnit}
\usage{
whodunnit(x = 1)
}
\arguments{
\item{x}{an integer value of how many rumours you want to generate}
}
\value{
x number of rumours returned
}
\description{
Whodunnit
}
\examples{
whodunnit()
whodunnit(5)
}
